requirements.txt
numpy
pandas
matplotlib
biopython
sklearn
hdbscan
notebook
seaborn
